+++ src/gmon.java
@@ -11,5 +11,10 @@
 // This class exists simply for the possibility that I want to add boot arguments
 public class gmon {
 
+    // TODO ASAP: license under open source license. Currently leaning towards Mozilla.
+    // TODO: Go through the entire project and add comments where needed to justify/explain choices.
+    // TODO: Fall the principles of clean code. If the code is clean, comments aren't usually necessary.
+    // TODO: Go through and refactor things that seem too messy.
+    // v2.0.0 TODO: Begin work on v2.0.0. v2 is going to focus on added more panels (VRAM usage, etc), clock speeds, and choosing panels to display.
     public static void main(String[] args) { new Main("gmon"); }
 }