+++ src/ui/GUI.java
@@ -12,25 +12,39 @@ import java.io.IOException;
 
 public class GUI extends JPanel {
     private final Statistics stats;
-    private final TemperaturePanel temps;
-    private final FanSpeedPanel fSpeeds;
-    private final TimePanel times;
+    private TemperaturePanel temps;
+    private FanSpeedPanel fSpeeds;
+    private TimePanel times;
+    private final Color bgColor;
+    private final Color fgColor;
 
+    // TODO: Menu bar with options to create a custom fan curve + JFrame and JPanels for it.
+    // TODO: Decouple any of the data processing from the xPanel classes so that it can run independent of the GUI.
+    // TODO: Add more colors set in main (maybe an array?) to be able to pass to the xPanels. Especially red for critical temps.
+    //
     public GUI(Statistics stats, Color bgColor, Color fgColor) {
         setLayout(new GridLayout(3,1));
         this.stats = stats;
+        this.bgColor = bgColor;
+        this.fgColor = fgColor;
 
+        InitTrackingPanels();
+    }
+
+    private void InitTrackingPanels() {
+        System.out.println("Initializing monitoring panels.");
         temps = new TemperaturePanel(stats, bgColor, fgColor);
+        fSpeeds = new FanSpeedPanel(stats, bgColor, fgColor);
+        times = new TimePanel(stats, bgColor, fgColor);
+
         temps.setBackground(bgColor);
         temps.setForeground(fgColor);
         add(temps);
 
-        fSpeeds = new FanSpeedPanel(stats, bgColor, fgColor);
         fSpeeds.setBackground(bgColor);
         fSpeeds.setForeground(fgColor);
         add(fSpeeds);
 
-        times = new TimePanel(stats, bgColor, fgColor);
         times.setBackground(bgColor);
         times.setForeground(fgColor);
         add(times);